Added method to retrieve map files
can get map files from either the jar or from an external maps folder in the install directory
might need work once we figure out how to create a game with a map to actually use the external map
(i.e. we would have to remove the portion of the path that is specific to a users machine and then on the client side, they could recreate the path .... assuming a requirement that all clients have prepopulated their folders for maps).

Corrected issue where tank was used in collision instead of tankLayer.